requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
asyncpg==0.30.0
overmind==0.1.31
anthropic>=0.40.0
airbyte-agent-github
